Address PR review: extract shared multi-monitor geometry
Move target-screen selection and frame clamping into
MultiMonitorWindowGeometry so main-window recovery no longer depends on
SettingsWindowPresenter. Unify off-screen recovery through recoveredFrame for
both constrainFrameRect and display-disconnect paths, narrow
applyOffscreenRecoveryIfNeeded to CmuxMainWindow, and derive a guaranteed
off-screen test frame from connected display bounds.
