+++ Sources/App/CmuxMainWindow.swift
@@ -282,7 +282,10 @@ final class CmuxMainWindow: NSWindow {
         ) {
             return frameRect
         }
-        return super.constrainFrameRect(frameRect, to: screen)
+        return Self.recoveredOffscreenFrame(
+            frameRect,
+            mouseLocation: NSEvent.mouseLocation
+        ) ?? super.constrainFrameRect(frameRect, to: screen)
     }
 
     /// Whether `proposedFrame` is reachable enough across `visibleFrames` that
@@ -385,10 +388,9 @@ extension CmuxMainWindow {
 
     /// Move a main window back onto a connected display when its frame would be
     /// stranded off-screen (e.g. the display it lived on was disconnected).
-    /// Uses the same target-screen selection as Settings multi-monitor recovery
-    /// (#5770). No-op when the frame already has a grabbable on-screen slice.
+    /// No-op when the frame already has a grabbable on-screen slice.
     static func applyOffscreenRecoveryIfNeeded(
-        _ window: NSWindow,
+        _ window: CmuxMainWindow,
         mouseLocation: NSPoint? = NSEvent.mouseLocation
     ) {
         guard !window.styleMask.contains(.fullScreen) else { return }
@@ -398,26 +400,30 @@ extension CmuxMainWindow {
             return
         }
 
-        let screens = NSScreen.screens.map { (frame: $0.frame, visibleFrame: $0.visibleFrame) }
-        let fallbackVisibleFrame = (NSScreen.main ?? NSScreen.screens.first)?.visibleFrame
-        guard let targetVisibleFrame = SettingsWindowPresenter.targetVisibleFrame(
-            windowFrame: window.frame,
-            screens: screens,
-            mouseLocation: mouseLocation,
-            fallbackVisibleFrame: fallbackVisibleFrame
-        ) else { return }
-
-        let minimumFrameSize = NSSize(
-            width: max(window.minSize.width, minimumContentSize.width),
-            height: max(window.minSize.height, minimumContentSize.height)
-        )
-        let recovered = SettingsWindowPresenter.clampedFrame(
-            window.frame,
-            minimumSize: minimumFrameSize,
-            into: targetVisibleFrame,
-            inset: 0
-        )
+        guard let recovered = recoveredOffscreenFrame(window.frame, mouseLocation: mouseLocation) else {
+            return
+        }
         guard recovered != window.frame else { return }
         window.setFrame(recovered, display: true, animate: false)
+    }
+
+    static func recoveredOffscreenFrame(
+        _ frame: NSRect,
+        mouseLocation: NSPoint?
+    ) -> NSRect? {
+        let screens = NSScreen.screens.map { (frame: $0.frame, visibleFrame: $0.visibleFrame) }
+        let fallbackVisibleFrame = (NSScreen.main ?? NSScreen.screens.first)?.visibleFrame
+        let minimumFrameSize = NSSize(
+            width: minimumContentSize.width,
+            height: minimumContentSize.height
+        )
+        return MultiMonitorWindowGeometry.recoveredFrame(
+            frame,
+            minimumSize: minimumFrameSize,
+            screens: screens,
+            mouseLocation: mouseLocation,
+            fallbackVisibleFrame: fallbackVisibleFrame,
+            inset: 0
+        )
     }
 }
