+++ Sources/AppDelegate.swift
@@ -18558,7 +18558,8 @@ extension AppDelegate {
 
     private func recoverMainWindowsAfterDisplayChange() {
         for window in mainWindowsForVisibilityController() where window.isVisible {
-            CmuxMainWindow.applyOffscreenRecoveryIfNeeded(window)
+            guard let mainWindow = window as? CmuxMainWindow else { continue }
+            CmuxMainWindow.applyOffscreenRecoveryIfNeeded(mainWindow)
         }
     }
 }
